Application controller provides all services to the interfaces; allows for code reuse for web and cli interfaces
