+++ src/main/java/ohtu/service/Tip.java
@@ -1,0 +1,51 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package ohtu.service;
+
+public class Tip {
+
+    private Integer id;
+    private String name;
+    private String type;
+
+    public Tip() {
+    }
+
+    public Tip(Integer id, String name, String type) {
+        this.id = id;
+        this.name = name;
+        this.type = type;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }    
+    
+    public String toString() {
+        return "id: " + this.id + " name: " + this.name + " type: " + this.type;
+    }
+    
+
+}
