+++ src/main/java/ohtu/web/Controllers.java
@@ -1,4 +1,4 @@
-package ohtu;
+package ohtu.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
