+++ src/main/java/ohtu/web/Main.java
@@ -1,4 +1,4 @@
-package ohtu;
+package ohtu.web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
