--- src/main/java/ohtucli/data_access/VinkkiDao.java
@@ -1,11 +1,0 @@
-
-package ohtucli.data_access;
-
-import java.util.List;
-import ohtucli.domain.Vinkki;
-
-public interface VinkkiDao {
-    List<Vinkki> listAll();
-    Vinkki findById(int id);
-    void add(Vinkki vinkki);
-}
