--- src/main/java/ohtucli/domain/Tip.java
@@ -1,33 +1,0 @@
-package ohtucli.domain;
-
-public class Tip {
-
-    private Integer id;
-    private String name;
-
-    public Tip(Integer id) {
-    }
-
-    public Tip(Integer id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-    
-
-}
