--- src/main/java/ohtucli/io/IO.java
@@ -1,8 +1,0 @@
-
-package ohtucli.io;
-
-public interface IO {
-    void print(String toPrint);
-    int readInt(String prompt);
-    String readLine(String prompt);
-}
